Decouple custom functions registration to JUSTContext class
