+++ JUST.net/JsonTransformer.cs
@@ -12,27 +12,6 @@ namespace JUST
 {
     public class JsonTransformer
     {
-        private static ConcurrentDictionary<string, MethodInfo> _customFunctions = new ConcurrentDictionary<string, MethodInfo>();
-
-        public static void RegisterCustomFunction(string assemblyName, string namespc, string methodName, string methodAlias = null)
-        {
-            var methodInfo = ReflectionHelper.SearchCustomFunction(assemblyName, namespc, methodName);
-            if (methodInfo == null)
-            {
-                throw new Exception("Unable to find specified method!");
-            }
-
-            if (!_customFunctions.TryAdd(methodAlias ?? methodName, methodInfo))
-            {
-                throw new Exception("Error while registering custom method!");
-            }
-        }
-
-        public static void ClearCustomFunctionRegistrations()
-        {
-            _customFunctions.Clear();
-        }
-
         public static string Transform(string transformerJson, string inputJson)
         {
             JToken result = null;
@@ -627,8 +606,9 @@ namespace JUST
                     output = ReflectionHelper.caller(null, "JUST.Transformer", functionName, new object[] { array, currentArrayElement, arguments[0] });
                 else if (functionName == "customfunction")
                     output = CallCustomFunction(parameters);
-                else if (_customFunctions.TryGetValue(functionName, out var methodInfo))
+                else if (JUSTContext.IsRegisteredCustomFunction(functionName))
                 {
+                    var methodInfo = JUSTContext.GetCustomMethod(functionName);
                     output = ReflectionHelper.InvokeCustomMethod(methodInfo, parameters, true);
                 }
                 else if (Regex.IsMatch(functionName, ReflectionHelper.EXTERNAL_ASSEMBLY_REGEX)){
