+++ Program.cs
@@ -208,17 +208,17 @@ namespace JUST.NET.Test
             Console.WriteLine("################################################################################################");
 
             transformer = File.ReadAllText("Examples/Transformer_registeredmethods.json");
-            JsonTransformer.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "StaticMethod", "External_StaticMethod");
-            JsonTransformer.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "StaticTypedParameters", "External_StaticTypedParameters");
-            JsonTransformer.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "InstanceMethod", "External_InstanceMethod");
-            JsonTransformer.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "TypedParameters", "External_TypedParameters");
-            JsonTransformer.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "NavigateTypedParameters", "External_NavigateTypedParameters");
+            JUSTContext.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "StaticMethod", "External_StaticMethod");
+            JUSTContext.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "StaticTypedParameters", "External_StaticTypedParameters");
+            JUSTContext.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "InstanceMethod", "External_InstanceMethod");
+            JUSTContext.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "TypedParameters", "External_TypedParameters");
+            JUSTContext.RegisterCustomFunction("ExternalMethods", "ExternalMethods.ExternalClass", "NavigateTypedParameters", "External_NavigateTypedParameters");
 
-            JsonTransformer.RegisterCustomFunction(null, "InternalMethods.InternalClass", "StaticMethod");
-            JsonTransformer.RegisterCustomFunction(null, "InternalMethods.InternalClass", "StaticTypedParameters");
-            JsonTransformer.RegisterCustomFunction(null, "InternalMethods.InternalClass", "InstanceMethod");
-            JsonTransformer.RegisterCustomFunction(null, "InternalMethods.InternalClass", "TypedParameters");
-            JsonTransformer.RegisterCustomFunction(null, "InternalMethods.InternalClass", "NavigateTypedParameters");
+            JUSTContext.RegisterCustomFunction(null, "InternalMethods.InternalClass", "StaticMethod");
+            JUSTContext.RegisterCustomFunction(null, "InternalMethods.InternalClass", "StaticTypedParameters");
+            JUSTContext.RegisterCustomFunction(null, "InternalMethods.InternalClass", "InstanceMethod");
+            JUSTContext.RegisterCustomFunction(null, "InternalMethods.InternalClass", "TypedParameters");
+            JUSTContext.RegisterCustomFunction(null, "InternalMethods.InternalClass", "NavigateTypedParameters");
             
             transformedString = JsonConvert.SerializeObject
                 (JsonTransformer.Transform(JObject.Parse(transformer), JObject.Parse(input)));
